pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "solr_mcp_server"
version = "1.0.0"
description = "An MCP Server for Apache Solr document search"
authors = [
    { name = "Martin Jochum", email = "mjochum64@gmail.com" }
]
readme = "README.md"
license = { file = "LICENSE" }
keywords = ["mcp", "solr", "search", "apache-solr", "document-search"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
]
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.0.0",
    "httpx>=0.24.1",
    "mcp[cli]>=1.6.0",
    "pydantic>=2.0.0",
    "pyjwt>=2.6.0",
    "passlib>=1.7.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black",
    "flake8",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "--strict-markers --tb=short"

[tool.flake8]
max-line-length = 88
exclude = [
    "__pycache__",
    ".git",
    ".venv",
    ".mypy_cache",
]

[tool.uv]
# MCP-specific configuration

[tool.coverage.run]
branch = true
source = ["src"]

[tool.coverage.report]
show_missing = true
skip_covered = true
